+++ pages/pollination-for-mentors.tsx
@@ -2,7 +2,7 @@ import type { NextPage } from "next"
 import Popper from "@popperjs/core"
 import Link from "next/link"
 import Head from "next/head"
-import Navbar from "../components/Navbar"
+import Navbar from "../components/layout/Navbar"
 import Footer from "../components/layout/Footer/footer"
 import Testimonials from "../components/sections/homepage/Testimonials"
 import React, { Component } from "react"
